+++ windows/timeseries_module.py
@@ -1,12 +1,4 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file 'ui_collection/time_series_module.ui'
-#
-# Created by: PyQt5 UI code generator 5.11.3
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore, QtGui, QtWidgets
+from PyQt5 import QtCore, QtWidgets
 from PyQt5.QtWidgets import QMainWindow
 
 
@@ -23,7 +15,8 @@ class Ui_MainWindow(object):
           USER INTERFACE OF TIME SERIES WINDOW 
         """""""""""""""""""""""""""""""""""""""
         MainWindow.setObjectName("MainWindow")
-        MainWindow.resize(615, 439)
+        MainWindow.setFixedHeight(439)
+        MainWindow.setFixedWidth(615)
         self.temp_window = MainWindow
         self.centralwidget = QtWidgets.QWidget(MainWindow)
         self.centralwidget.setObjectName("centralwidget")
@@ -48,7 +41,7 @@ class Ui_MainWindow(object):
         self.btn_modify.setObjectName("btn_modify")
         self.scrollArea_timeseries = QtWidgets.QScrollArea(self.centralwidget)
         self.scrollArea_timeseries.setGeometry(QtCore.QRect(50, 90, 329, 101))
-        self.scrollArea_timeseries.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOn)
+        self.scrollArea_timeseries.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAsNeeded)
         self.scrollArea_timeseries.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarAsNeeded)
         self.scrollArea_timeseries.setWidgetResizable(True)
         self.scrollArea_timeseries.setObjectName("scrollArea_timeseries")
@@ -61,7 +54,7 @@ class Ui_MainWindow(object):
         self.lbl_timeseries.setObjectName("lbl_timeseries")
         self.scrollArea_products = QtWidgets.QScrollArea(self.centralwidget)
         self.scrollArea_products.setGeometry(QtCore.QRect(50, 280, 329, 101))
-        self.scrollArea_products.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOn)
+        self.scrollArea_products.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAsNeeded)
         self.scrollArea_products.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarAsNeeded)
         self.scrollArea_products.setWidgetResizable(True)
         self.scrollArea_products.setObjectName("scrollArea_products")
@@ -106,7 +99,19 @@ class Ui_MainWindow(object):
         self.btn_bind.clicked.connect(self.btn_bind_clicked)
         self.btn_unbind.clicked.connect(self.btn_unbind_clicked)
 
-        self.products_list_view()
+        self.btn_delete.setDisabled(True)
+        self.btn_bind.setDisabled(True)
+        self.btn_unbind.setDisabled(True)
+        self.btn_modify.setDisabled(True)
+        self.btn_export.setDisabled(True)
+        self.btn_unbind.setDisabled(True)
+
+        self.txt_products.textChanged.connect(self.productfilterClicked)
+        self.txt_timeseries.textChanged.connect(self.timeseriesfilterClicked)
+
+        self.load_products_list()
+        self.selected_timeseries = None
+        self.selected_product = None
 
         self.retranslateUi(MainWindow)
         QtCore.QMetaObject.connectSlotsByName(MainWindow)
@@ -130,6 +135,7 @@ class Ui_MainWindow(object):
         from windows.timeseries_addNew import TimeSeriesAddNewWindow
         self.timeseries_addnew_win = TimeSeriesAddNewWindow(self.temp_window)
         self.timeseries_addnew_win.show()
+        self.temp_window.hide()
 
     def btn_return_clicked(self):
         self.temp_window.hide()
@@ -154,6 +160,7 @@ class Ui_MainWindow(object):
             output_file.close()
 
     def btn_modify_clicked(self):
+        import datetime
         result = [x.row() for x in self.timeseries_listWidget.selectedIndexes()]
         if result:
             timeseries_info = self.timeseries_info
@@ -166,35 +173,51 @@ class Ui_MainWindow(object):
             self.modify_win.ui.combo_analysetype.setCurrentText(timeseries_info["analyse_type"])
             self.modify_win.ui.txt_emptydata.setText(str(timeseries_info["empty_data"]))
             self.modify_win.ui.txt_seriesdata.setText(str(timeseries_info["series_data"]))
-            # self.modify_win.ui.txt_endtime.setTime(timeseries_info["end_time"])
-            # self.modify_win.ui.txt_starttime.setTime(timeseries_info["start_time"])
+            end_time = datetime.datetime.fromtimestamp(timeseries_info["end_time"]/1000).strftime('%Y-%m-%d %H:%M:%S.%f')
+            start_time = datetime.datetime.fromtimestamp(timeseries_info["start_time"]/1000).strftime('%Y-%m-%d %H:%M:%S.%f')
+            self.modify_win.ui.txt_endtime.setText(str(end_time)[:-7])
+            self.modify_win.ui.txt_starttime.setText(str(start_time)[:-7])
             self.modify_win.ui.txt_name.setText(timeseries_info["name"])
+            self.modify_win.ui.source_file = timeseries_info["source_file"]
             self.modify_win.ui.modify = True
             self.modify_win.show()
+            self.temp_window.hide()
 
     def btn_delete_clicked(self):
-        self.delete_timeseries(self.selected_timeseriesId)
-        # self.productId_list.remove(product_code)
-        self.timeseries_list.remove(self.selected_timeseries)
-        self.timeseriesId_list.remove(self.selected_timeseriesId)
-        self.timeseries_list_view()
+        buttonReply = QtWidgets.QMessageBox.question(self.temp_window, "Message",
+                                                     'Do you really want to delete this product ?',
+                                                     QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No,
+                                                     QtWidgets.QMessageBox.No)
+        if buttonReply == QtWidgets.QMessageBox.Yes:
+            self.delete_timeseries(self.selected_timeseriesId)
+            self.timeseries_list.remove(self.selected_timeseries)
+            self.timeseriesId_list.remove(self.selected_timeseriesId)
+            self.timeseries_list_view()
 
     def btn_bind_clicked(self):
-        from windows.database_util import DatabaseConnect
+        from utils.database_utils import DatabaseConnect
         db = DatabaseConnect()
         updated = db.update_timeseries_productId(self.selected_timeseriesId, self.selected_productId)
         if updated:
             print("Successfully bind the product with time series !!!")
+            self.btn_bind.setDisabled(True)
+            self.btn_unbind.setDisabled(True)
+            self.timeseries_list_view()
+            self.load_products_list()
 
     def btn_unbind_clicked(self):
-        from windows.database_util import DatabaseConnect
+        from utils.database_utils import DatabaseConnect
         db = DatabaseConnect()
         updated = db.update_timeseries_productId(self.selected_timeseriesId, "NULL")
         if updated:
             print("Successfully unbind the product from time series !!!")
+            self.btn_bind.setDisabled(True)
+            self.btn_unbind.setDisabled(True)
+            self.timeseries_list_view()
+            self.load_products_list()
 
     def timeseries_list_view(self):
-        from windows.database_util import DatabaseConnect
+        from utils.database_utils import DatabaseConnect
         db = DatabaseConnect()
         self.timeseries_list, self.timeseriesId_list = db.get_timeserieses()
         self.timeseries_listWidget = QtWidgets.QListWidget()
@@ -212,31 +235,57 @@ class Ui_MainWindow(object):
     def timeseries_list_item_event(self, item):
         self.selected_timeseries = item.text()
         self.selected_timeseriesId = item.data(1)
+        self.btn_export.setEnabled(True)
+        self.btn_delete.setEnabled(True)
+        self.btn_modify.setEnabled(True)
         print(self.selected_timeseriesId)
         self.get_timeseries_details()
+        if self.selected_timeseries and self.selected_product:
+            if self.timeseries_info["product_ID"] == self.selected_productId:
+                self.btn_unbind.setEnabled(True)
+                self.btn_bind.setDisabled(True)
+            elif not self.timeseries_info["product_ID"] or self.timeseries_info["product_ID"] == "":
+                self.btn_bind.setEnabled(True)
+                self.btn_unbind.setDisabled(True)
+            else:
+                self.btn_bind.setDisabled(True)
+                self.btn_unbind.setDisabled(True)
 
     def get_timeseries_details(self):
-        from windows.database_util import DatabaseConnect
+        from utils.database_utils import DatabaseConnect
         db = DatabaseConnect()
         self.timeseries_info = db.get_timeseries_details(self.selected_timeseriesId)
         return self.timeseries_info
 
     def delete_timeseries(self, id):
-        from windows.database_util import DatabaseConnect
+        from utils.database_utils import DatabaseConnect
         db = DatabaseConnect()
         deleted = db.delete_timeseries(id=id)
         if deleted:
             print("Successfully deleted the time series !!!")
 
+    def timeseriesfilterClicked(self):
+        filter_text = str(self.txt_timeseries.text()).lower()
+        self.timeseries_listWidget.clear()
+        index = 0
+        for item in self.timeseries_list:
+            if item.lower().startswith(filter_text.lower()):
+                listitem = QtWidgets.QListWidgetItem()
+                listitem.setText(item)
+                listitem.setData(1, self.timeseriesId_list[index])
+                self.timeseries_listWidget.addItem(listitem)
+            index += 1
+        self.scrollArea_timeseries.setWidget(self.timeseries_listWidget)
+
     """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
                                       PRODUCT PART
     """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
     def get_products(self):
-        from windows.database_util import DatabaseConnect
+        from utils.database_utils import DatabaseConnect
         db = DatabaseConnect()
         self.products_list, self.productId_list = db.get_products()
 
-    def products_list_view(self):
+    def load_products_list(self):
         self.listWidget_product = QtWidgets.QListWidget()
         self.get_products()
         index = 0
@@ -250,10 +299,32 @@ class Ui_MainWindow(object):
         self.scrollArea_products.setWidget(self.listWidget_product)
 
     def list_item_event(self, item):
-        print(repr(item.text()))
         self.selected_product = item.text()
         self.selected_productId = item.data(1)
-        print(self.selected_productId)
+        if self.selected_timeseries and self.selected_product:
+            if self.timeseries_info["product_ID"] == self.selected_productId:
+                self.btn_unbind.setEnabled(True)
+                self.btn_bind.setDisabled(True)
+            elif not self.timeseries_info["product_ID"] or self.timeseries_info["product_ID"] == "":
+                self.btn_bind.setEnabled(True)
+                self.btn_unbind.setDisabled(True)
+            else:
+                self.btn_bind.setDisabled(True)
+                self.btn_unbind.setDisabled(True)
+
+
+    def productfilterClicked(self):
+        filter_text = str(self.txt_products.text()).lower()
+        self.listWidget_product.clear()
+        index = 0
+        for item in self.products_list:
+            if item.lower().startswith(filter_text.lower()):
+                listitem = QtWidgets.QListWidgetItem()
+                listitem.setText(item)
+                listitem.setData(1, self.productId_list[index])
+                self.listWidget_product.addItem(listitem)
+            index += 1
+        self.scrollArea_products.setWidget(self.listWidget_product)
     """"""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
